add amcreate

+++ controllers/repo.go
@@ -6,6 +6,7 @@ import (
 	log "github.com/astaxie/beego/logs"
 	"github.com/gzsunrun/ansible-manager/core/config"
 	"github.com/gzsunrun/ansible-manager/core/function"
+	"github.com/gzsunrun/ansible-manager/core/template"
 	"github.com/gzsunrun/ansible-manager/core/orm"
 	"github.com/gzsunrun/ansible-manager/core/storage"
 	"github.com/satori/go.uuid"
@@ -50,11 +51,17 @@ func (c *RepoController) Create() {
 	}
 	defer os.Remove(config.Cfg.Common.WorkPath + "/" + repoPath)
 	defer os.RemoveAll(config.Cfg.Common.WorkPath + "/" + repoPath + "_dir")
-	err = function.ReadVars(config.Cfg.Common.WorkPath+"/"+repoPath, &repo)
+	// err = function.ReadVars(config.Cfg.Common.WorkPath+"/"+repoPath, &repo)
+	// if err != nil {
+	// 	c.SetResult(err, nil, 400)
+	// 	return
+	// }
+	tpls,err:=template.ReadVars(config.Cfg.Common.WorkPath+"/"+repoPath,repoPath)
 	if err != nil {
 		c.SetResult(err, nil, 400)
 		return
 	}
+
 	if c.GetString("repo_name") != "" {
 		repo.Name = c.GetString("repo_name")
 	}
@@ -72,7 +79,7 @@ func (c *RepoController) Create() {
 			c.SetResult(err, nil, 400)
 			return
 		}
-	}
+	} 
 	repoParse := storage.StorageParse{
 		LocalPath:  config.Cfg.Common.WorkPath + "/" + repoPath,
 		RemotePath: repoPath,
@@ -82,11 +89,16 @@ func (c *RepoController) Create() {
 		c.SetResult(err, nil, 400)
 		return
 	}
-	err = orm.CreateRepo(repo)
+	err = orm.CreateRepos(tpls)
 	if err != nil {
 		c.SetResult(err, nil, 400)
 		return
 	}
+	// err = orm.CreateRepo(repo)
+	// if err != nil {
+	// 	c.SetResult(err, nil, 400)
+	// 	return
+	// }
 	c.SetResult(nil, nil, 204)
 }
 
@@ -104,14 +116,16 @@ func (c *RepoController) Delete() {
 		c.SetResult(err, nil, 400)
 		return
 	}
-	logoParse := storage.StorageParse{
-		RemotePath: repo.Path + ".png",
+	if !orm.GetRepoByPath(repo.Path){
+		logoParse := storage.StorageParse{
+			RemotePath: repo.Path + ".png",
+		}
+		repoParse := storage.StorageParse{
+			RemotePath: repo.Path,
+		}
+		storage.Storage.Delete(&logoParse)
+		storage.Storage.Delete(&repoParse)
 	}
-	repoParse := storage.StorageParse{
-		RemotePath: repo.Path,
-	}
-	storage.Storage.Delete(&logoParse)
-	storage.Storage.Delete(&repoParse)
 	c.SetResult(nil, nil, 204)
 }
